EC2: Support encrypted volumes without storing a key
This commit adds an option ‘deployment.ec2.storeKeysOnRootDisk’ that,
if set to false, prevents LUKS encryption keys of EBS volumes from
being stored on the root disk.  Instead, "charon deploy" uploads the
encryption keys to /run/ebs-keys, which is a tmpfs lost after a
reboot.

This increases security: if an attacker gains access to the root disk
of a (stopped) instance, he won't be able to decrypt the data disks.
However, it prevents unattended reboots (e.g. as a result of a kernel
panic).  After a reboot, the init-luks Upstart job waits until the
keys appear in /run/ebs-keys, then proceeds with the rest of the boot
process.  Re-uploading the keys can be triggered by running "charon
deploy".

TODO: add a command "charon send-keys" that just sends the encryption
keys to the instance without doing a redeployment?  Maybe "charon
deploy --create-only" is sufficient.

Fixes #51.