hetzner: Install partitioner and run in rescue.
We already had the ability to boot into the rescue system, now we're actually
doing something, namely: partitioning. The partitioner itself spits out a Nix
expression with the created filesystems which can be used to provide a physical
base (fileSystem/swapDevices) specification of the machine. All we need to do
now is to deploy into the chroot and we're finished with the basics.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>