Factor out logging into nixops.logger.
This should make it easier, to extend logging features, such as for example to
use pythons "logging" module. The new logger module is pretty much basic and
just resembles the functionality we had before but in *one* place.

Also, the base class ResourceState now just contains wrapper methods to
self.logger, so we don't break existing functionality.

So, for the new logger: The goal is to have a base logger (Logger), which is
used to log non-machine-specific stuff. Whenever we want to have a Logger for a
specific machine, we use base_logger.get_logger_for() to create and return a new
logger just for the machine.

Using this approach has some benefits:

 - We don't need to carry around logging prefixes everywhere in the code.
 - It easily allows to use context managers for start/end-like logging.
 - Extension is easier.

And of course, we can do things in the category "neat", such progressbars or
this:

machine1> Doing something.......
machine2> Doing something...........done.
machine3> Doing something....

Where all lines don't get mixed up and interrupted by other lines and thus stay
consistent (in the terminal _and_ the logfile).

The name "logger" was choosen to avoid clashes with Python's native logging
module.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>