logger: Fix coloring on first machine deployment.
When the MachineState was initalized, it got a machine logger instance
based on the value of the current self.index. Unfortunately for new
machines, the index is assigned after __init__ and the first log line,
so we're missing the index in the logger and thus coloring.

This is now temporarily fixed at the place where the machine gets
assigned an index. But optimally would be if we would have some way to
do coloring solely based on the machines name and without getting
duplicate colors on <= 7 machines.

So, until we're able to figure this out, we have a new method called
register_index(), which is there so it's literally crying to be factored
out :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>