ssh_util: Add reset() method to reset SSHMaster.
This is needed when for example flags or passwords were changed in the
MachineState. We're using weakrefs here, to ensure that even when the
master is left referenced somewhere in the wild, it will still be the
only reference left within SSH.

Probably the best method to solve this is using context managers, but
that's entirely out of the scope of this branch and will be something to
work on afterwards.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>