tests/hetzner: Add and integrate target exprs.
This adds the Nix expressions needed to deploy the two target machines.
Both have different partition schemes, so we can test later on if the
partitioning was applied correctly.

Of course, this adds everything that is needed to use the network
expressions on the target machines. The only thing that's missing now is
the coordinator itself.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>