nix_expr: Add a function to merge node constants.
This adds the enclose_node() function, which is used to merge a prefix
and/or suffix into another node while being agnostic on whether it is a
RawValue or a Container. For the latter we need to also take care of
properly merging the value into the inline_variant of the correspending
Container.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>