nix_expr: Improve and rename merge_dicts().
The function is now called nixmerge() and is able to merge not only
dicts but also lists and maybe more someday.

One limitation is that it cannot merge nested lists and or lists
containing dicts, because the merge is done with a union of two sets and
all items within a set need to be hashable.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>