hetzner: Fix local_commands in _gen_network_spec.
This appends a newline at the end of the string, to allow for easier
concatenation of networking.localCommands and avoid confusion in case
localCommands suddenly stop working as expected.

Also, when pretty-printing it's also easier to print the string in a
multi-line variant rather than either completely inline or just with
ugly quoting.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>