hetzner: Restructure generation of physical info.
This now encodes fs_info and hw_info with their surrounding braces
and/or the function head verbatim as they come from nixos-hardware-scan
and nixpart. The networking info now is encoded in JSON and thus the
machine state reflects this now properly.

The reason of this change is to not clutter up the physical spec any
more than necessary and the second reason is because we can now embed
the network info as the top-level of the backend's physical spec and
import fs_info and hw_info using nix2py.

Now, anyone who is using the Hetzner backend already needs to convert
his/her machine state to the new format. I've written a small update
script for that, which is available at:

https://gist.github.com/aszlig/6449773

Usage is something like this (assuming your current working directory is
a clone of the Gist):

nix-build && result/bin/nixops-update-hetzner-netinfo

Of course, if you already have sqlite3 in your PYTHONPATH you can also
directly run "python update.py".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>