nix_expr: Remove parsing functionality of nix2py.
This should get rid of the parsing routines and the values fed into
nix2py are now just used as multiple raw values. The reason for this is
because it's actually not necessary anymore to merge, as we are now
using the module system.

Even though with the parser the pretty printed result looks really way
better than without:

https://gist.github.com/aszlig/850c88f54a479bc9866d

So, rest in peace little parser, as I don't think anyone else would
*demand* prettier output of the physical spec ;-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>