nix/keys: Correctly merge/convert deployment.keys.
So far we only used the apply function in order to convert keys as
string values into the new format. But if you have several keys defined,
mixing old and new format, the evaluation will fail because the apply
function is only evaluated once everything is merged.

So, we now use the same function we're using for apply to merge the
consecutive option values.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>